requirements.txt
requests==2.32.3
terminaltable==3.1.10
python-dotenv==1.0.1 